Move ChainMetadata to `tari_common_types` crate to fix wallet_ffi dependency
There is an existing issue when building ‘wallet_ffi’ for Android or iOS where the `croaring` library used in `tari_core`. To avoid this issue we have been using feature flags to exclude the parts of `tari_core` that uses `croaring`. However, now that we need the ChainMetadata struct to deserialise the chain metadata  message in the wallet made it difficult to use the feature flags for this purpose.  

To fix this issue this PR moves the ChainMetadata and BlockHash types to a new crate called `tari_common_types` so we can use this type without needing to include the `croaring` dependancy.
