+++ applications/tari_app_grpc/Cargo.toml
@@ -8,6 +8,7 @@ version = "0.6.1"
 edition = "2018"
 
 [dependencies]
+tari_common_types = { version = "^0.1", path = "../../base_layer/common_types"}
 tari_core = {  path = "../../base_layer/core"}
 tari_crypto = { version = "^0.8" }
