+++ base_layer/core/Cargo.toml
@@ -19,6 +19,7 @@ avx2 = ["tari_crypto/avx2"]
 
 [dependencies]
 tari_common = { version = "^0.2", path = "../../common"}
+tari_common_types = { version = "^0.1", path = "../../base_layer/common_types"}
 tari_comms = { version = "^0.2", path = "../../comms"}
 tari_comms_dht = { version = "^0.2.1", path = "../../comms/dht"}
 tari_comms_rpc_macros = { version = "^0.1", path = "../../comms/rpc_macros"}
