+++ base_layer/core/src/base_node/state_machine_service/states/events_and_states.rs
@@ -20,22 +20,20 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use crate::{
-    base_node::state_machine_service::states::{
-        BlockSyncInfo,
-        BlockSyncStrategy,
-        HeaderSync,
-        HorizonStateSync,
-        Listening,
-        ListeningInfo,
-        Shutdown,
-        Starting,
-        SyncPeers,
-        Waiting,
-    },
-    chain_storage::ChainMetadata,
+use crate::base_node::state_machine_service::states::{
+    BlockSyncInfo,
+    BlockSyncStrategy,
+    HeaderSync,
+    HorizonStateSync,
+    Listening,
+    ListeningInfo,
+    Shutdown,
+    Starting,
+    SyncPeers,
+    Waiting,
 };
 use std::fmt::{Display, Error, Formatter};
+use tari_common_types::chain_metadata::ChainMetadata;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum BaseNodeState {
