+++ base_layer/core/src/blocks/mod.rs
@@ -28,6 +28,6 @@ mod new_blockheader_template;
 pub mod genesis_block;
 
 pub use block::{Block, BlockBuilder, BlockValidationError, NewBlock};
-pub use blockheader::{BlockHash, BlockHeader, BlockHeaderValidationError};
+pub use blockheader::{BlockHeader, BlockHeaderValidationError};
 pub use new_block_template::NewBlockTemplate;
 pub use new_blockheader_template::NewBlockHeaderTemplate;
