+++ base_layer/wallet/src/base_node_service/handle.rs
@@ -24,7 +24,8 @@ use super::{error::BaseNodeServiceError, service::BaseNodeState};
 use futures::{stream::Fuse, StreamExt};
 use std::sync::Arc;
 use tari_comms::peer_manager::Peer;
-use tari_core::chain_storage::ChainMetadata;
+
+use tari_common_types::chain_metadata::ChainMetadata;
 use tari_service_framework::reply_channel::SenderService;
 use tokio::sync::broadcast;
 use tower::Service;
